Frist Actions
The tweets slice of the state in the store will be modified by actions that go through the tweets reducer. The users slice of the state in the store will be modified by actions that go through the users reducer. And, similarly, the authedUser portion of the state in the store will be modified by actions that go through the authedUser reducer.

We will create our reducers for this initial data action in the next commit. It is a good practise to have the Actions implemented first

Did you know that there are other side effect management libraries besides redux-thunk that help you make async requests like
Redux-Saga  https://hackernoon.com/moving-api-requests-to-redux-saga-21780f49cbc8
Redux-axios  https://github.com/svrcekmichal/redux-axios-middleware

If you want to learn more about the differences between these libraries, check out this article.
https://decembersoft.com/posts/what-is-the-right-way-to-do-asynchronous-operations-in-redux/