When we walked through the architecture of our app, we saw that the App Component will contain every other component. If we load the initial data (by dispatching the handleInitialData() action creator) from the App component, then no matter which route our users goes to, they’ll see all of the correct data.
Make sure to go through the excellent examples that are provided in the linked documentation to gain a deeper understanding of Redux.
https://github.com/reduxjs/react-redux/tree/master/docs/api

REMINDER: the handleInitialData method is an action creator. Typically action creators returns objects that the dispatch method will execute but due to the thunk we imported from redux-thunk, we are able to also invoke action creators that are functions

AT this point we can see the log with our state three initlalized with data correclty

I had to fix a minor bug on handleInitialData()
https://stackoverflow.com/questions/42299594/await-is-a-reserved-word-error-inside-async-function