Liking a Tweet - Redux Reducer
When liking a tweet (or unliking a tweet), the state for that specific tweet needs to change - either the tweet's like property (which, if you remember, is an array and will contain the names of the users that have liked the tweet) will need to change to include the user that clicked it (if they're liking the tweet) or the user's name will need to be removed from the array (if they're unliking the tweet).

So we need to update the reducer to handle these changes.

- Inside the reducer/tweets.js, we imported the TOGGLE_LIKE action type
- We added a new case to our reducer for if the case is TOOGLE_LIKE, how do we want to change the shape of the tweets slice of our state tree

Remember in your reducers you will use the spread operator, .filert(), .concat(), .map(), object.assign(), .sort() a lot. Basically any functionality that does not mutate data. This means, that you get thesame result even as long a you invoke them with thesame inputs