React Router
BrowserRouter listens for changes in the URL and makes sure that the correct screen shows up when the URL changes.

Now we have successfully built out individual screens inside our application:
- We want to add react-router so that we can dynamically change what component is being rendered based on the URL and We want add a Nav bar so that user can go between Dashboard or NewTweet component. Tihs was Nav.js and App.js
- We want to redirect the user back to the Dashboard if they are posting a new tweet or remain on the TweetPage if they are replying to a tweet. This was implemented in NewtTweet.js
- We wanted the user to be able to click on these different tweets to be taken to the tweet page. This was implemented in Tweet.js

Note: What if you wanted to pass state to the new route? Instead of passing a string to Links to prop, you can pass it an object like this:

<Link to={{
 pathname: '/courses',
 search: '?sort=name',
 hash: '#the-hash',
 state: { fromDashboard: true }
}}>
 Courses
</Link>